requirements.txt
Flask
flask-cors
llama-index
llama-index-embeddings-google-genai
llama-index-llms-google-genai
google-generativeai
faiss-cpu
PyMuPDF
python-dotenv